Restructure module manually.
